+++ src/components/customer/CustomerDetailComponent.vue
